% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{build_path}
\alias{build_path}
\title{Function to create full endpoint path}
\usage{
build_path(s)
}
\arguments{
\item{s}{One of 'basic' (for 'property/basicprofile'), 'detail'
(for 'property/detail'), 'address' (for 'property/address'),
or 'sales' (for 'sale/snapshot')}
}
\value{
path Full endpoint path (endpoint/service)
}
\description{
This helper function creates the full endpoint path (e.g.,
'endpoint/property/snapshot) to pass to function attom_api()
}
