% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{build_query}
\alias{build_query}
\title{Function to build base query}
\usage{
build_query(s, ...)
}
\arguments{
\item{s}{One of 'basic' (for 'property/basicprofile'), 'detail'
(for 'property/detail'), 'address' (for 'property/address'),
or 'sales' (for 'sale/snapshot')}

\item{...}{Other named query parameters}
}
\value{
query List of query parameters to pass to attom_api()
}
\description{
This helper function builds a basic list of query parameters that
is common across queries
}
