% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{create_queries}
\alias{create_queries}
\title{Function to create list of lists of addresses from data frame}
\usage{
create_queries(d)
}
\arguments{
\item{d}{A data frame such that each row is an address}
}
\value{
queries A list of lists of queries
}
\description{
Function to create list of lists of addresses from data frame
}
