% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{parse_basic}
\alias{parse_basic}
\title{Function to post-process responses for basic query}
\usage{
parse_basic(resp, el = "property")
}
\arguments{
\item{resp}{A single response from try_query()}

\item{el}{Element from response to post-process}
}
\value{
d A data frame of parsed property characteristics
}
\description{
Function to post-process responses for basic query
}
