% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{parse_list}
\alias{parse_list}
\title{Function to post-process list of responses}
\usage{
parse_list(resps, el = "property")
}
\arguments{
\item{resps}{List of responses (e.g., from search_list())}
}
\value{
parsed_list List of data frames with response property data
}
\description{
Given a list of responses, extract the response data.frame that
contains the property information (NB: for now, only parses responses
from the basic query).
}
