% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{parse_sales}
\alias{parse_sales}
\title{Function to post-process responses for sales query}
\usage{
parse_sales(resp, el = "property")
}
\arguments{
\item{resp}{A single response from try_query()}

\item{el}{Element from response to post-process}
}
\value{
out A data frame of parsed properties associated with query
}
\description{
Function to post-process responses for sales query
}
