% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{print.attom_api}
\alias{print.attom_api}
\title{Print GET response}
\usage{
\method{print}{attom_api}(x, ...)
}
\arguments{
\item{x}{An object of class \code{attom_api}}
}
\description{
Adds print method to object of class \code{attom_api} (borrowed from
httr vignette)
}
