% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{search_list}
\alias{search_list}
\title{Function to iterate through list of queries (list of addresses)}
\usage{
search_list(queries, apikey, s, ...)
}
\arguments{
\item{queries}{A list of lists, where each list is a set of query
parameters to pass to attom_api()}

\item{apikey}{The user's ATTOM API Key}

\item{s}{One of 'basic' (for 'property/basicprofile'), 'detail'
(for 'property/detail'), 'address' (for 'property/address'),
or 'sales' (for 'sale/snapshot')}

\item{...}{Other named query parameters (passed to build_query())}
}
\value{
responses A list of responses of class \code{attom_api}, for each
item in the list queries
}
\description{
Given a list of queries, iteratively call attom_api() by building
and passing a query
}
