% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{set_ua}
\alias{set_ua}
\title{function to optionally set user agent globally}
\usage{
set_ua(a)
}
\arguments{
\item{a}{A string to pass to httr::user_agent}
}
\description{
function to optionally set user agent globally
}
