% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{try_query}
\alias{try_query}
\title{Error handling for queries}
\usage{
try_query(...)
}
\arguments{
\item{...}{Other named query parameters (passed to attom_api())}
}
\value{
result Result of query with error handling
}
\description{
Error handling for queries
}
