% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attom_api.R
\name{update_query}
\alias{update_query}
\title{Function to update query}
\usage{
update_query(query, l)
}
\arguments{
\item{query}{A list of named query parameters}

\item{l}{A list of named parameters to append to query}
}
\value{
query An updated list of named query parameters
}
\description{
This helper function updates a list of query parameters by
appending new parameters to the query
}
